+++ custom_components/elastic/es_version.py
@@ -1,0 +1,38 @@
+"""Maintains information about the verion of Elasticsearch"""
+
+class ElasticsearchVersion:
+    """Maintains information about the verion of Elasticsearch"""
+
+    def __init__(self, client):
+        self._client = client
+        self._version_number_str = None
+        self.major = None
+        self.minor = None
+        self.build_flavor = None
+
+    async def async_init(self):
+        """I/O bound init"""
+        version = (await self._client.info())["version"]
+        version_number_parts = version["number"].split(".")
+        self._version_number_str = version["number"]
+        self.major = int(version_number_parts[0])
+        self.minor = int(version_number_parts[1])
+        self.build_flavor = version["build_flavor"]
+
+    def is_supported_version(self):
+        """Determines if this version of ES is supported by this component"""
+        return self.major == 7 or (
+            self.major == 6 and self.minor >= 7
+        )
+
+    def is_oss_distribution(self):
+        """Determines if this is the OSS distribution"""
+        return self.build_flavor == 'oss'
+
+    def is_default_distribution(self):
+        """Determines if this is the default distribution"""
+        return self.build_flavor == 'default'
+
+    def to_string(self):
+        """Returns a string representation of the current ES version"""
+        return self._version_number_str
