+++ custom_components/elastic/logger.py
@@ -1,0 +1,5 @@
+""" Component Logger """
+
+import logging
+
+LOGGER = logging.getLogger("custom_components.elastic")
